// packages/api/src/codex-protocol/v2/UserInput.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import { type TextElement } from "./TextElement";

export type UserInput =
  | {
      type: "text";
      text: string;
      /**
       * UI-defined spans within `text` used to render or persist special elements.
       */
      text_elements: Array<TextElement>;
    }
  | { type: "image"; url: string }
  | { type: "localImage"; path: string }
  | { type: "skill"; name: string; path: string }
  | { type: "mention"; name: string; path: string };